+++ src/main/java/se/sics/hop/metadata/hdfs/dal/INodeAttributesDataAccess.java
@@ -10,9 +10,9 @@ import se.sics.hop.exception.StorageException;
  */
 public interface INodeAttributesDataAccess<T> extends EntityDataAccess {
 
-  T findAttributesByPk(long inodeId) throws StorageException;
+  T findAttributesByPk(Integer inodeId) throws StorageException;
   
-  Collection<T> findAttributesByPkList(Collection<Long> inodeIds) throws StorageException;
+  Collection<T> findAttributesByPkList(Collection<Integer> inodeIds) throws StorageException;
 
   void prepare(Collection<T> modified, Collection<T> removed) throws StorageException;
 }
